the most rudimentary of weapon plates are ready to be put into a container. added a neat little pathfinding display debug thing

+++ core/src/com/deeper/than/weapons/WeaponParams.java
@@ -15,7 +15,7 @@ public class WeaponParams{
 	public float critChance;
 	public float rechargeSpeed;
 	public float baseMonetaryCost;
-	public float powerCost;
+	public int powerCost;
 	public WeaponQualities quality;
 	public WeaponMakers maker;
